+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - On X11, add mappings for numpad comma, numpad enter, numlock and pause.
+- On Web, use the canvas size on window creation when a canvas is provided.
 
 # 0.26.0 (2021-12-01)